+++ docs/changelog.rst
@@ -17,6 +17,8 @@ Dependencies
 
 - Python >= 3.9 is now required. Python 3.7 and 3.8 are no longer supported.
 
+- GStreamer >= 1.18.0 is now required.
+
 
 v3.4.1 (2022-12-07)
 ===================