+++ docs/changelog.rst
@@ -9,6 +9,12 @@ This changelog is used to track all major changes to Mopidy.
 For older releases, see :ref:`history`.
 
 
+v3.4.2 (UNRELEASED)
+===================
+
+- Dev: Added "tox -e ci", to allow easy CI check before "git push".
+
+
 v3.4.1 (2022-12-07)
 ===================
 
@@ -49,7 +55,7 @@ v3.3.0 (2022-04-29)
   fields, with the exception of 'any'. Deprecated field 'track' with the
   goal of removing it in the next major release, use 'track_name' instead.
   Backends should support both `track` and `track_name` until they require
-  a version of Mopidy where `track` has been removed. 
+  a version of Mopidy where `track` has been removed.
   (Fixes: :issue:`1900`, PR: :issue:`1899`)
 
 - Core: Add ``musicbrainz_albumid``, ``musicbrainz_artistid``,